N1ebieski/vs-code-extension#104 Support for basePath
